+++ src/providers/provider.interface.ts
@@ -1,13 +1,7 @@
-import type { Artist, OAuthTokens, Track } from "../types/index.js";
+import type { OAuthTokens, Track } from "../types/index.js";
 
 export interface MusicProvider {
   getTopTracks: (accessToken: string) => Promise<Track[]>;
-  getTopArtists: (accessToken: string) => Promise<Artist[]>;
-  searchByGenre: (
-    accessToken: string,
-    genre: string,
-    limit?: number,
-  ) => Promise<Track[]>;
   createOrUpdatePlaylist: (
     accessToken: string,
     userId: string,